+++ README.md
@@ -1,4 +1,4 @@
-# :jp: DomainValidatorJp
+# :jp: DomainValidator::Jp
 
 <a href="https://rubygems.org/gems/domain-validator-jp" title="npm"><img src="https://img.shields.io/gem/v/domain-validator-jp.svg"></a>
 <a href="https://travis-ci.org/kimromi/domain-validator-jp" title="travis"><img src="https://img.shields.io/travis/kimromi/domain-validator-jp.svg"></a>
@@ -9,11 +9,11 @@ domain name validator includes JP domain
 ## Usage
 
 ```rb
-DomainValidatorJp.new('example.jp').valid?        # => true
-DomainValidatorJp.new('日本語.jp').valid?         # => true
+DomainValidator::Jp.new('example.jp').valid?        # => true
+DomainValidator::Jp.new('日本語.jp').valid?         # => true
 
-DomainValidatorJp.new('①②③.jp').valid?            # => false
-DomainValidatorJp.new('example.wrongtld').valid?  # => false
+DomainValidator::Jp.new('①②③.jp').valid?            # => false
+DomainValidator::Jp.new('example.wrongtld').valid?  # => false
 ```
 
 ## Installation
@@ -21,7 +21,7 @@ DomainValidatorJp.new('example.wrongtld').valid?  # => false
 Add this line to your application's Gemfile:
 
 ```ruby
-gem 'domain-validator-jp'
+gem 'domain_validator-jp'
 ```
 
 And then execute:
@@ -30,7 +30,7 @@ And then execute:
 
 Or install it yourself as:
 
-    $ gem install domain-validator-jp
+    $ gem install domain_validator-jp
 
 ## Development
