+++ lib/domain_validator-jp.rb
@@ -1,0 +1,7 @@
+require "domain_validator/jp/version"
+require "domain_validator/jp"
+
+module DomainValidator
+  class Jp
+  end
+end
