+++ lib/domain_validator/jp/version.rb
@@ -1,0 +1,5 @@
+module DomainValidator
+  class Jp
+    VERSION = "0.1.1"
+  end
+end
