--- lib/domain_validator_jp/version.rb
@@ -1,3 +1,0 @@
-class DomainValidatorJp
-  VERSION = "0.1.1"
-end
